requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.5.0
    # via pydantic
anyio==4.0.0
    # via httpcore
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.2.0
    # via requests
h11==0.14.0
    # via httpcore
httpcore==0.18.0
    # via httpx
httpx==0.25.0
    # via letsbuilda-pypi
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
letsbuilda-pypi==5.0.0
    # via -r requirements.in
pydantic==2.4.2
    # via pydantic-settings
pydantic-core==2.10.1
    # via pydantic
pydantic-settings==2.0.3
    # via -r requirements.in
python-dotenv==1.0.0
    # via
    #   -r requirements.in
    #   pydantic-settings
requests==2.31.0
    # via -r requirements.in
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
typing-extensions==4.8.0
    # via
    #   pydantic
    #   pydantic-core
urllib3==2.0.5
    # via requests
xmltodict==0.13.0
    # via letsbuilda-pypi
